Eliminate "signed and unsigned type in conditional expression" warnings   [-Wsign-compare]
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@18997 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/lcd_time.c
@@ -511,7 +511,7 @@ hash_engineID(const u_char * engineID, u_int engineID_len)
     SNMP_FREE(context);
     memset(buf, 0, SNMP_MAXBUF);
 
-    return (rval < 0) ? rval : (additive % ETIMELIST_SIZE);
+    return (rval < 0) ? rval : (int)(additive % ETIMELIST_SIZE);
 
 }                               /* end hash_engineID() */
